requirements.txt
Django
djangorestframework
prometheus-client
requests
